split MarkdownRenderer
