+++ src/components/SideBar/CustomViewer/SidebarItem.tsx
@@ -47,7 +47,9 @@ import { PageButton } from 'components/NavBar/PageButton';
 import Description from '@mui/icons-material/Description';
 import dayjs from 'dayjs';
 import { addRecentPage } from 'features/currentSlices';
-import { ArrowDropDown, ArrowRight, DescriptionOutlined } from '@mui/icons-material';
+import ArrowDropDown from '@mui/icons-material/ArrowDropDown';
+import ArrowRight from '@mui/icons-material/ArrowRight';
+import DescriptionOutlined from '@mui/icons-material/DescriptionOutlined';
 import { Theme } from 'features/settingSlices';
 import { findColor } from 'utils/document';
 import { SideBarItemList } from '../SidebarItemList';
